backend/requirements.txt
fastapi==0.72.0
python-dotenv==0.19.2
sqlalchemy==1.4.29
alembic==1.7.5
python-slugify==5.0.2 